% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotSLM.r
\name{boxplotSLM}
\alias{boxplotSLM}
\title{Boxplot of sample L-moments (boxplotSLM)}
\usage{
boxplotSLM(data, groups, dataName = NULL, showPoints = TRUE,
  groupCol = NULL, ...)
}
\arguments{
\item{data}{data matrix}

\item{groups}{a character vector indicating sample group membership}

\item{dataName}{character vector indicating name of dataset.}

\item{showPoints}{show data points}

\item{groupCol}{group color}

\item{...}{arguments passed to boxplots}
}
\description{
Boxplot of sample L-moments (boxplotSLM)
}

