% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/computeDvals.R
\name{computeDvals}
\alias{computeDvals}
\title{Compute d-values.}
\usage{
computeDvals(t3, t4, plot = FALSE, span = 0.5)
}
\arguments{
\item{t3}{a vector containing L-skew estimates for each gene.}

\item{t4}{a vector containing L-kurt estimates for each gene.}

\item{plot}{indicating whether to show intermediary plots (Default=FALSE).}

\item{span}{the lowess smoother span}
}
\description{
Compute d-values.
}

