% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filterCounts.r
\name{filterCounts}
\alias{filterCounts}
\title{Filter out genes with low counts}
\usage{
filterCounts(counts, lib.size = NULL, thresh = 1, minSamples = 2)
}
\arguments{
\item{counts}{raw counts matrix}

\item{lib.size}{library size (defualt = NULL)}

\item{thresh}{cpm threshold (defualt =1)}

\item{minSamples}{number of samples required to exceed threshold (default=2)}
}
\description{
Filter out genes with low counts
}

