% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitShape.R
\name{fitShape}
\alias{fitShape}
\title{Sample L-moments ratios.}
\usage{
fitShape(x, nLmom = 4)
}
\arguments{
\item{x}{data vector / matrix.}

\item{nLmom}{the number of l-moments to compute (Default = 4).}

\item{samplesByCols}{do columns represent samples ? (Default = TRUE)}
}
\description{
This function computes the sample l-moments ratios for a given
dataset (vector / matrix / data.frame).
}

