% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitShape_internals.R
\name{Lmoments}
\alias{Lmoments}
\title{Compute Lmoments (Code lifted from Lmoments package).}
\usage{
Lmoments(data, rmax)
}
\arguments{
\item{data}{matrix or dataframe}

\item{rmax}{number of lmoments}
}
\description{
Compute Lmoments (Code lifted from Lmoments package).
}

