% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitShape_internals.R
\name{Lmoments_calc}
\alias{Lmoments_calc}
\title{Compute Lmoments (Code lifted from Lmoments package).}
\usage{
Lmoments_calc(data, rmax = 4)
}
\arguments{
\item{data}{matrix or dataframe}

\item{rmax}{number of lmoments}
}
\description{
Compute Lmoments (Code lifted from Lmoments package).
}

