% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lweights.r
\name{lweights}
\alias{lweights}
\title{Compute weights for each expression based on l-scale}
\usage{
lweights(exprs, groups, span = 0.5, plot = FALSE)
}
\arguments{
\item{exprs}{normalized expression on log scale}

\item{groups}{character vector indicating the group membership of each sample}

\item{span}{span for lowess (default=0.5)}

\item{plot}{show fit ? (default=FALSE)}
}
\description{
Compute weights for each expression based on l-scale
}

