% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shapeFit.R
\name{shapeFit}
\alias{shapeFit}
\title{This function computes L-skew and L-kurt estimates.}
\usage{
shapeFit(x = TRUE)
}
\arguments{
\item{x}{Do you love cats :) Defaults to TRUE.}
}
\description{
This function computes L-skew and L-kurt estimates.
}

