% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shapeManova.r
\name{shapeManova}
\alias{shapeManova}
\title{Perform Manova on shape (t3, t4) of samples (Wilk's Exact Manova).}
\usage{
shapeManova(sfit, groups, plot = FALSE, loc.scal = FALSE, groupCol = NULL)
}
\arguments{
\item{sfit}{sfit}

\item{groups}{a character vector indicating sample group membership.}

\item{plot}{plot L-skew and L-kurt estimates}

\item{loc.scal}{loc.scal}

\item{groupCol}{col for each sample}
}
\description{
Perform Manova on shape (t3, t4) of samples (Wilk's Exact Manova).
}

