% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotSLM.r
\name{boxplotSLM}
\alias{boxplotSLM}
\title{Boxplot of sample L-moments (boxplotSLM)}
\usage{
boxplotSLM(data, groups, data.name = NULL, show.points = TRUE,
  group.col = NULL, ...)
}
\arguments{
\item{data}{data matrix}

\item{groups}{a character vector indicating sample group membership}

\item{data.name}{character vector indicating name of dataset.}

\item{show.points}{show.points}

\item{group.col}{group.col}

\item{...}{arguments passed to boxplots}
}
\description{
Boxplot of sample L-moments (boxplotSLM)
}

