% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shapeFit.R
\name{shapeFit}
\alias{shapeFit}
\title{Sample L-moments ratios.}
\usage{
shapeFit(x, nLmom = 4, samplesByCols = TRUE)
}
\arguments{
\item{x}{data vector / matrix.}

\item{nLmom}{the number of l-moments to compute (Default = 4).}

\item{samplesByCols}{do columns represent samples ? (Default = TRUE)}
}
\description{
This function computes the sample l-moments ratios for a given
dataset (vector / matrix / data.frame).
}

