% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shapeManova.r
\name{shapeManova}
\alias{shapeManova}
\title{Perform Manova on shape (t3, t4) of samples (Wilk's Exact Manova).}
\usage{
shapeManova(data, groups, nPerm = 500, lrats = FALSE, plot = FALSE,
  groupCol = NULL)
}
\arguments{
\item{groups}{a character vector indicating sample group membership}

\item{nPerm}{number of permutations for non-parametric test}

\item{lrats}{lrats}

\item{plot}{plot}

\item{groupCol}{group color}

\item{sfit}{output from fitShape()}
}
\description{
Perform Manova on shape (t3, t4) of samples (Wilk's Exact Manova).
}

