% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shapeManova.r
\name{shapeManova}
\alias{shapeManova}
\title{Perform Manova on shape (t3, t4) of samples (Wilk's Exact Manova).}
\usage{
shapeManova(sfit, groups, nPerm = 500)
}
\arguments{
\item{sfit}{output from fitShape()}

\item{groups}{a character vector indicating sample group membership}

\item{nPerm}{number of permutations for non-parametric test}
}
\description{
Perform Manova on shape (t3, t4) of samples (Wilk's Exact Manova).
}

