% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shapeManova.r
\name{shapeManova}
\alias{shapeManova}
\title{Perform Manova on shape (t3, t4) of samples (Wilk's Exact Manova).}
\usage{
shapeManova(lrats, groups)
}
\arguments{
\item{lrats}{a (2 by number of samples) matrix containing (t3, t3) for each sample.}

\item{groups}{a character vector indicating sample group membership.}
}
\description{
Perform Manova on shape (t3, t4) of samples (Wilk's Exact Manova).
}

