% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shapeManova.r
\name{shapeManova}
\alias{shapeManova}
\title{Perform Manova on shape (t3, t4) of samples (Wilk's Exact Manova).}
\usage{
shapeManova(exprs, groups, plot = TRUE, groupCol = NULL)
}
\arguments{
\item{groups}{a character vector indicating sample group membership.}

\item{plot}{plot L-skew and L-kurt estimates}

\item{groupCol}{col for each sample}

\item{lrats}{a (2 by number of samples) matrix containing (t3, t3) for each sample.}
}
\description{
Perform Manova on shape (t3, t4) of samples (Wilk's Exact Manova).
}

